hindsite-backend/requirements.txt
fastapi 
uvicorn 
python-dotenv 
cohere 
psycopg2-binary 
pgvector 
sqlalchemy